Implemented keybinds
Added /bind, /unbind, and /unbindall, to manage custom keybinds with Console.
InterpretLua and InterpretCommand no longer automatically execute the chunk and return its value- instead, the chunk is returned (or the error string as the second argument, if unsuccessful), and the caller handles executing the chunk. This allows keybinds to cache functions of specified command or chunk type.
Keybinds are also saved between sessions.
Added a function to check whether the console window is currently active